feat: wired up frontend

+++ Dtos/EntryResponse.cs
@@ -8,7 +8,7 @@ public class EntryResponse
     public int UserId { get; set; }
     public long GameId { get; set; }
     public EntryStatus Status { get; set; }
-    public float? Rating { get; set; }
+    public byte? Rating { get; set; }
     public string? Notes { get; set; }
     public DateTime CreatedAt { get; set; }
     public DateTime UpdatedAt { get; set; }
